greenswap_backend/requirements-prod.txt
# Production requirements
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-redis==5.4.0
Pillow==10.1.0
django-cors-headers==4.3.1
channels==4.0.0
channels-redis==4.1.0
celery==5.3.4
redis==5.0.1
django-filter==23.5
openai==1.3.8
python-dotenv==1.0.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
django-storages==1.14.2
boto3==1.34.0
sentry-sdk==1.38.0